refactor: move NumberPad into shared components
