+++ src/systems/funding/components/DepositScreen.tsx
@@ -18,10 +18,10 @@ import { TIMING_CONFIGS } from '@/components/animations/animationConfigs';
 import RainbowCoinIcon from '@/components/coin-icon/RainbowCoinIcon';
 import Page from '@/components/layout/Page';
 import { Navbar } from '@/components/navbar/Navbar';
+import { NumberPad } from '@/components/number-pad/NumberPad';
 import { DecoyScrollView } from '@/components/sheet/DecoyScrollView';
 import { Box, useColorMode, useForegroundColor } from '@/design-system';
 import { GasSpeed } from '@/features/gas/types/gasSpeed';
-import { NumberPad } from '@/features/perps/components/NumberPad/NumberPad';
 import { SheetHandle } from '@/features/perps/components/SheetHandle';
 import { SLIDER_MAX, type SliderColors } from '@/features/perps/components/Slider/Slider';
 import { SliderWithLabels } from '@/features/perps/components/Slider/SliderWithLabels';
