+++ src/systems/funding/components/WithdrawalScreen.tsx
@@ -9,10 +9,10 @@ import { AccountImage } from '@/components/AccountImage';
 import RainbowCoinIcon from '@/components/coin-icon/RainbowCoinIcon';
 import Page from '@/components/layout/Page';
 import { Navbar } from '@/components/navbar/Navbar';
+import { InputValueCaret } from '@/components/number-pad/InputValueCaret';
+import { NumberPad } from '@/components/number-pad/NumberPad';
 import { AnimatedText, Box, Inline, Inset, Text, TextIcon, useColorMode, useForegroundColor } from '@/design-system';
 import { palettes } from '@/design-system/color/palettes';
-import { InputValueCaret } from '@/features/perps/components/InputValueCaret';
-import { NumberPad } from '@/features/perps/components/NumberPad/NumberPad';
 import { PerpsSwapButton } from '@/features/perps/components/PerpsSwapButton';
 import { PerpsTextSkeleton } from '@/features/perps/components/PerpsTextSkeleton';
 import { SheetHandle } from '@/features/perps/components/SheetHandle';
