+++ src/systems/funding/components/deposit/DepositAmountInput.tsx
@@ -8,9 +8,9 @@ import { getColorValueForThemeWorklet } from '@/__swaps__/utils/swaps';
 import { AnimatedTextIcon } from '@/components/AnimatedComponents/AnimatedTextIcon';
 import { GestureHandlerButton } from '@/components/buttons/GestureHandlerButton';
 import { ImgixImage } from '@/components/images';
+import { InputValueCaret } from '@/components/number-pad/InputValueCaret';
 import { AnimatedText, Bleed, Box, Column, Columns, Separator, Stack, Text, useColorMode } from '@/design-system';
 import { type TextColor } from '@/design-system/color/palettes';
-import { InputValueCaret } from '@/features/perps/components/InputValueCaret';
 import { PerpsTextSkeleton } from '@/features/perps/components/PerpsTextSkeleton';
 import * as i18n from '@/languages';
 import { DepositQuoteStatus } from '@/systems/funding/types';
