+++ src/systems/funding/hooks/useDepositController.ts
@@ -5,7 +5,7 @@ import { useDebouncedCallback } from 'use-debounce';
 
 import { type ExtendedAnimatedAssetWithColors } from '@/__swaps__/types/assets';
 import { SPRING_CONFIGS } from '@/components/animations/animationConfigs';
-import { type NumberPadField } from '@/features/perps/components/NumberPad/NumberPadKey';
+import { type NumberPadField } from '@/components/number-pad/NumberPadKey';
 import { SLIDER_MAX } from '@/features/perps/components/Slider/Slider';
 import {
   divWorklet,
